dnsdist: Optionally keep all data in LRU cache
To be able to keep all data in cache, and not delete
expired, we need another mechanism to evict from cache;
and LRU is the most useful here, as the actually used elements
will stay in the cache.

MaybeLruCache is either behaving like a regular map, if isLRU is false,
or also has a linked list+map for constant-time putFront

One of the changes is also that finding in cache can now require a write lock.

Signed-off-by: Karel Bilek <kb@karelbilek.com>